MR_2023_11_02

+++ 2023/AdventOfCode2023/AdventOfCode2023/Models/Day11/GalaxyPair.cs
@@ -2,7 +2,7 @@
 
 public class GalaxyPair
 {
-    public int ManhattanDistance;
+    public long ManhattanDistance;
     private readonly Galaxy _galaxy1;
     private readonly Galaxy _galaxy2;
